move transfer status to bridge package from stores package
