+++ packages/bridge/package.json
@@ -40,7 +40,6 @@
     "@keplr-wallet/unit": "0.10.24-ibc.go.v7.hot.fix",
     "@osmosis-labs/math": "^5.1.0",
     "@osmosis-labs/pools": "^5.1.0",
-    "@osmosis-labs/stores": "^5.1.0",
     "@osmosis-labs/types": "^1.0.0",
     "@osmosis-labs/utils": "^1.0.0",
     "cachified": "^3.5.4",
